Added a hoc for protected pages.
