+++ src/components/loading.tsx
@@ -1,0 +1,11 @@
+import LoadingSpinner from "@/components/UI/loading-spinner";
+
+const Loading = () => {
+  return (
+    <div className="bg-white flex h-screen w-full items-center justify-center">
+      <LoadingSpinner />
+    </div>
+  );
+};
+
+export default Loading;
